вроде работает поиск билетов в обе стороны

+++ src/main/java/com/example/prac/data/res/RouteDTO.java
@@ -8,7 +8,7 @@ import java.util.List;
 public class RouteDTO {
     private String departureCity;
     private String arrivalCity;
-    private int totalHours;
+    private double totalHours;
     private double totalPrice;
     private List<TicketDTO> tickets;
 }
